move the mesh_resource_provider_system into a SystemNode of the render graph
