+++ crates/bevy_render/src/render_graph/nodes/mod.rs
@@ -1,4 +1,5 @@
 mod camera_node;
+mod mesh_node;
 mod pass_node;
 mod render_resources_node;
 mod shared_buffers_node;
@@ -7,6 +8,7 @@ mod window_swapchain_node;
 mod window_texture_node;
 
 pub use camera_node::*;
+pub use mesh_node::*;
 pub use pass_node::*;
 pub use render_resources_node::*;
 pub use shared_buffers_node::*;
